+++ src/pages/Home/components/RoadMapSection/roadMapStyle.js
@@ -1,0 +1,13 @@
+const style = {
+  roadmap: {
+    marginBottom: 100
+  },
+  title: {
+    fontSize: '100px',
+    fontWeight: 700,
+    textAlign: 'center',
+    color: '#fff'
+  }
+}
+
+export default style
