--- src/pages/Home/components/roadMapStyle.js
@@ -1,13 +1,0 @@
-const style = {
-  roadmap: {
-    marginBottom: 100
-  },
-  title: {
-    fontSize: '100px',
-    fontWeight: 700,
-    textAlign: 'center',
-    color: '#fff'
-  }
-}
-
-export default style
