--- src/pages/Home/components/roadMapV2Style.js
@@ -1,62 +1,0 @@
-const style = {
-  roadmap: {
-    marginBottom: 100
-  },
-  title: {
-    fontSize: '5rem',
-    fontWeight: 700,
-    textAlign: 'center',
-    color: '#fff'
-  },
-  bottomButton: {
-    color: '#A68EFE',
-    border: '0.5rem solid #A68EFE',
-    fontSize: '2.5rem',
-    textAlign: 'center',
-    fontWeight: 700,
-    borderRadius: '8rem',
-    width: '18.25rem',
-    margin: '0 auto',
-    letterSpacing: '0.01em',
-    lineHeight: 2.7
-  },
-  text: {
-    fontSize: '1.375rem',
-    textAlign: 'center',
-    borderRadius: '12rem',
-    margin: '0 auto',
-    marginBottom: '5rem'
-  },
-  textInner: {
-    display: 'inline-block',
-    marginBottom: 0,
-    fontWeight: 700,
-    fontSize: '1.1875rem',
-    lineHeight: '120%',
-    letterSpacing: '0.01em'
-  },
-  textInnerTitle: {
-    padding: '0 10%'
-  },
-  dataItemContainer: {
-    padding: '2rem 2.3rem 0.6rem'
-  },
-  itemTitle: {
-    margin: 0,
-    display: 'inline-block',
-    padding: '0 0.5rem',
-    borderRadius: '2.5rem',
-    position: 'absolute',
-    transform: 'translate(-50%,-50%)',
-    letterSpacing: '0.01em',
-    lineHeight: '1.875rem',
-    fontWeight: 'bold',
-    fontStyle: 'italic',
-    fontSize: '0.875rem'
-  },
-  dot: {
-    marginTop: '0.5rem'
-  }
-}
-
-export default style
